added platformbuilder, works well

+++ werescrewed/src/com/blindtigergames/werescrewed/platforms/RoomPlatform.java
@@ -20,10 +20,10 @@ import com.blindtigergames.werescrewed.screens.GameScreen;
 
 public class RoomPlatform extends Platform {
 	protected boolean rotate = false;
-	protected int tileHeight, tileWidth;
+	protected float tileHeight, tileWidth;
 
-	public RoomPlatform( String n, Vector2 pos, Texture tex, int width,
-			int height, World world ) {
+	public RoomPlatform( String n, Vector2 pos, Texture tex, float width,
+			float height, World world ) {
 		super( n, pos, tex, world );
 		this.tileHeight = height;
 		this.tileWidth = width;
@@ -32,7 +32,7 @@ public class RoomPlatform extends Platform {
 		constructRoomBody( pos.x, pos.y, width, height );
 	}
 
-	public void constructRoomBody( float x, float y, int width, int height ) {
+	public void constructRoomBody( float x, float y, float width, float height ) {
 		PolygonShape ps = new PolygonShape( );
 		FixtureDef fd = new FixtureDef( );
 		fd.density = 1f;
